tests/test_generate_data.py
#!/usr/bin/env python3
"""
test_generate_data.py
---------------------
Unit tests for the data generation script.
"""

import os
import shutil
import tempfile
import unittest
from unittest.mock import patch, MagicMock

import pandas as pd
from pyspark.sql import SparkSession
from delta import DeltaTable

# Import the functions from generate_data.py
# Assuming generate_data.py has a main function or can be imported
# For simplicity, we'll mock and test key components


class TestDataGeneration(unittest.TestCase):

    def setUp(self):
        """Set up test fixtures."""
        self.temp_dir = tempfile.mkdtemp()
        self.delta_path = os.path.join(self.temp_dir, "delta-lake", "customer_transactions")

    def tearDown(self):
        """Clean up test fixtures."""
        shutil.rmtree(self.temp_dir, ignore_errors=True)

    def test_delta_table_creation(self):
        """Test that Delta table is created with correct schema."""
        # This would require running the actual script
        # For now, we'll test the components

        # Mock Spark session
        with patch('generate_data.build_spark') as mock_build_spark:
            mock_spark = MagicMock()
            mock_build_spark.return_value = mock_spark

            # Mock DataFrame operations
            mock_df = MagicMock()
            mock_spark.createDataFrame.return_value = mock_df
            mock_df.write.format.return_value = mock_df
            mock_df.mode.return_value = mock_df
            mock_df.save.return_value = None

            # Import and run the generation (this is pseudo-code)
            # from scripts.generate_data import main
            # main()

            # Verify Delta table creation was called
            mock_df.write.format.assert_called_with("delta")
            mock_df.mode.assert_called_with("overwrite")
            mock_df.save.assert_called_with(self.delta_path)

    def test_data_schema(self):
        """Test that generated data has the correct schema."""
        # Test the expected schema structure
        expected_columns = [
            'transaction_id', 'customer_id', 'customer_name', 'customer_email',
            'merchant_id', 'merchant_name', 'amount', 'currency', 'transaction_date',
            'payment_method', 'status', 'category', 'location'
        ]

        # This would check if the DataFrame has these columns
        # For now, just assert the list exists
        self.assertEqual(len(expected_columns), 13)
        self.assertIn('transaction_id', expected_columns)
        self.assertIn('amount', expected_columns)

    def test_data_validation(self):
        """Test data quality checks."""
        # Test that amounts are positive
        # Test that dates are within range
        # Test that IDs are unique

        # Mock data for testing
        test_data = {
            'transaction_id': ['txn_001', 'txn_002'],
            'amount': [100.50, 250.75],
            'transaction_date': [pd.Timestamp('2024-01-01'), pd.Timestamp('2024-01-02')]
        }
        df = pd.DataFrame(test_data)

        # Assertions
        self.assertTrue(all(df['amount'] > 0))
        self.assertEqual(len(df['transaction_id'].unique()), len(df))
        self.assertTrue(all(df['transaction_date'] <= pd.Timestamp.now()))


if __name__ == '__main__':
    unittest.main()