tests/requirements.txt
pytest>=7.0.0
pytest-mock>=3.10.0
apache-airflow>=2.8.0
pyspark>=3.4.0
delta-spark>=2.4.0
pandas>=1.5.0
faker>=15.0.0